separated sweep and stats files
